C: Use `uint32_t` for `length` member in `hb_string_T`  struct (#640)
In line with the lexer/parser we are going to limit the string length to
the range of `uint32_t` which is sufficient to hold 4 GB large strings.

Resolves #636

+++ src/include/util/hb_string.h
@@ -3,10 +3,11 @@
 
 #include <stdbool.h>
 #include <stddef.h>
+#include <stdint.h>
 
 typedef struct HB_STRING_STRUCT {
   char* data;
-  size_t length;
+  uint32_t length;
 } hb_string_T;
 
 hb_string_T hb_string_from_c_string(const char* null_terminated_c_string);
